* network.c (setup_network): Cast argument to isspace() to int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13372 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/kdc/network.c
@@ -545,7 +545,7 @@ setup_network(prog)
     for (i=0; i<kdc_numrealms; i++) {
 	cp = kdc_realmlist[i]->realm_ports;
 	while (cp && *cp) {
-	    if (*cp == ',' || isspace(*cp)) {
+	    if (*cp == ',' || isspace((int) *cp)) {
 		cp++;
 		continue;
 	    }
